extract common jsonrpc tests into separate factory
