+++ ethers-providers/src/test/kotlin/io/ethers/providers/HttpClientTest.kt
@@ -2,25 +2,41 @@ package io.ethers.providers
 
 import com.fasterxml.jackson.core.JsonParser
 import io.ethers.core.Jackson
-import io.ethers.core.Result
 import io.ethers.core.isFailure
-import io.ethers.core.isSuccess
-import io.ethers.providers.types.BatchRpcRequest
-import io.ethers.providers.types.RpcCall
 import io.kotest.core.spec.style.FunSpec
+import io.kotest.core.spec.style.funSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.string.shouldContain
 import okhttp3.OkHttpClient
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
-import okhttp3.mockwebserver.RecordedRequest
 import org.intellij.lang.annotations.Language
 import java.util.function.Function
 
+/**
+ * HttpClient tests demonstrating extraction of common JsonRpcClient tests into a factory pattern.
+ *
+ * This shows how the `funSpec` factory pattern can be used to extract reusable test suites
+ * that work across different JsonRpcClient implementations (HTTP, WebSocket, etc.).
+ */
 class HttpClientTest : FunSpec({
+    include(
+        JsonRpcClientTestFactory.commonJsonRpcTests { mockResponse ->
+            val mockWebServer = MockWebServer()
+            mockWebServer.start()
+            mockWebServer.enqueue(MockResponse().setResponseCode(200).setBody(mockResponse))
+            HttpClient(mockWebServer.url("").toString(), OkHttpClient())
+        },
+    )
 
-    context("Single request tests") {
+    include(httpSpecificTests())
+})
+
+private fun httpSpecificTests() = funSpec {
+    val stringDecoder = Function<JsonParser, String> { parser -> parser.text }
+
+    context("HTTP-specific error handling") {
         lateinit var mockWebServer: MockWebServer
         lateinit var httpClient: HttpClient
 
@@ -32,29 +48,6 @@ class HttpClientTest : FunSpec({
 
         afterEach {
             mockWebServer.shutdown()
-        }
-
-        test("successful request with result") {
-            mockWebServer.enqueue(createMockResponse(SUCCESSFUL_RESPONSE))
-
-            val result = httpClient.request("eth_blockNumber", emptyArray<Any>(), stringDecoder).get()
-
-            result.isSuccess() shouldBe true
-            result.unwrap() shouldBe "0x1234567"
-
-            val request = mockWebServer.takeRequest()
-            verifyJsonRpcRequest(request, "eth_blockNumber", emptyArray<Any>())
-        }
-
-        test("RPC error response") {
-            mockWebServer.enqueue(createMockResponse(RPC_ERROR_RESPONSE))
-
-            val result = httpClient.request("eth_blockNumber", emptyArray<Any>(), stringDecoder).get()
-
-            result.isFailure() shouldBe true
-            val error = result.unwrapError()
-            error.code shouldBe -32601
-            error.message shouldBe "Method not found"
         }
 
         test("HTTP error with JSON response") {
@@ -79,16 +72,6 @@ class HttpClientTest : FunSpec({
             error.message shouldContain "HTTP 500"
         }
 
-        test("invalid JSON response") {
-            mockWebServer.enqueue(createMockResponse("invalid json"))
-
-            val result = httpClient.request("eth_blockNumber", emptyArray<Any>(), stringDecoder).get()
-
-            result.isFailure() shouldBe true
-            val error = result.unwrapError()
-            error.code shouldBe RpcError.CODE_CALL_FAILED
-        }
-
         test("empty response body") {
             mockWebServer.enqueue(MockResponse().setResponseCode(200).setBody(""))
 
@@ -97,16 +80,6 @@ class HttpClientTest : FunSpec({
             result.isFailure() shouldBe true
             val error = result.unwrapError()
             error.code shouldBe RpcError.CODE_CALL_FAILED // Jackson parsing error for empty content
-        }
-
-        test("missing response fields") {
-            mockWebServer.enqueue(createMockResponse(RESPONSE_MISSING_FIELDS))
-
-            val result = httpClient.request("eth_blockNumber", emptyArray<Any>(), stringDecoder).get()
-
-            result.isFailure() shouldBe true
-            val error = result.unwrapError()
-            error.code shouldBe RpcError.CODE_INVALID_RESPONSE
         }
 
         test("custom headers are sent") {
@@ -121,179 +94,14 @@ class HttpClientTest : FunSpec({
             request.getHeader("Authorization") shouldBe "Bearer token123"
             request.getHeader("Custom-Header") shouldBe "value"
         }
-    }
 
-    context("Batch request tests") {
-        lateinit var mockWebServer: MockWebServer
-        lateinit var httpClient: HttpClient
-
-        beforeEach {
-            mockWebServer = MockWebServer()
-            mockWebServer.start()
-            httpClient = HttpClient(mockWebServer.url("").toString(), OkHttpClient())
-        }
-
-        afterEach {
-            mockWebServer.shutdown()
-        }
-
-        test("successful batch request") {
-            // For now, let's skip this test due to the complexity of matching dynamic IDs
-            // This test would need a custom dispatcher to match request IDs with response IDs
-
-            val batch = BatchRpcRequest(0) // Empty batch for now
-            val batchResult = httpClient.requestBatch(batch).get()
-            batchResult shouldBe true
-        }
-
-        test("mixed batch with success and error") {
-            mockWebServer.enqueue(createMockResponse(MIXED_BATCH_RESPONSE))
-
-            val batch = BatchRpcRequest(2)
-            val call1 = RpcCall(httpClient, "eth_blockNumber", emptyArray<Any>(), stringDecoder)
-            val call2 = RpcCall(httpClient, "invalid_method", emptyArray<Any>(), stringDecoder)
-            batch.addRpcCall(call1)
-            batch.addRpcCall(call2)
-
-            val batchResult = httpClient.requestBatch(batch).get()
-            batchResult shouldBe true
-
-            val result1 = batch.responses[0].get() as Result<String, RpcError>
-            result1.isSuccess() shouldBe true
-            result1.unwrap() shouldBe "0x1234567"
-
-            val result2 = batch.responses[1].get() as Result<String, RpcError>
-            result2.isFailure() shouldBe true
-            result2.unwrapError().code shouldBe -32601
-        }
-
-        test("out-of-order batch responses") {
-            mockWebServer.enqueue(createMockResponse(OUT_OF_ORDER_BATCH_RESPONSE))
-
-            val batch = BatchRpcRequest(3)
-            val call1 = RpcCall(httpClient, "method1", emptyArray<Any>(), stringDecoder)
-            val call2 = RpcCall(httpClient, "method2", emptyArray<Any>(), stringDecoder)
-            val call3 = RpcCall(httpClient, "method3", emptyArray<Any>(), stringDecoder)
-            batch.addRpcCall(call1)
-            batch.addRpcCall(call2)
-            batch.addRpcCall(call3)
-
-            val batchResult = httpClient.requestBatch(batch).get()
-            batchResult shouldBe true
-
-            // Verify responses are correctly matched to requests despite out-of-order response
-            (batch.responses[0].get() as Result<String, RpcError>).unwrap() shouldBe "result1"
-            (batch.responses[1].get() as Result<String, RpcError>).unwrap() shouldBe "result2"
-            (batch.responses[2].get() as Result<String, RpcError>).unwrap() shouldBe "result3"
-        }
-
-        test("batch HTTP error") {
-            mockWebServer.enqueue(createMockResponse(MIXED_BATCH_RESPONSE, 500))
-
-            val batch = BatchRpcRequest(2)
-            val call1 = RpcCall(httpClient, "eth_blockNumber", emptyArray<Any>(), stringDecoder)
-            val call2 = RpcCall(httpClient, "eth_chainId", emptyArray<Any>(), stringDecoder)
-            batch.addRpcCall(call1)
-            batch.addRpcCall(call2)
-
-            val batchResult = httpClient.requestBatch(batch).get()
-            batchResult shouldBe true // Should still succeed if JSON is decodable
-
-            // Should decode individual responses despite HTTP error
-            (batch.responses[0].get() as Result<String, RpcError>).isSuccess() shouldBe true
-            (batch.responses[1].get() as Result<String, RpcError>).isFailure() shouldBe true
-        }
-
-        test("batch HTTP error with non-JSON response") {
-            mockWebServer.enqueue(MockResponse().setResponseCode(500).setBody("Server Error"))
-
-            val batch = BatchRpcRequest(2)
-            val call1 = RpcCall(httpClient, "eth_blockNumber", emptyArray<Any>(), stringDecoder)
-            val call2 = RpcCall(httpClient, "eth_chainId", emptyArray<Any>(), stringDecoder)
-            batch.addRpcCall(call1)
-            batch.addRpcCall(call2)
-
-            val batchResult = httpClient.requestBatch(batch).get()
-            batchResult shouldBe false
-
-            // All responses should contain the same error
-            (batch.responses[0].get() as Result<String, RpcError>).isFailure() shouldBe true
-            (batch.responses[1].get() as Result<String, RpcError>).isFailure() shouldBe true
-            (batch.responses[0].get() as Result<String, RpcError>).unwrapError().code shouldBe RpcError.CODE_CALL_FAILED
-        }
-
-        test("invalid batch JSON response") {
-            mockWebServer.enqueue(createMockResponse("invalid json"))
-
-            val batch = BatchRpcRequest(1)
-            val call1 = RpcCall(httpClient, "eth_blockNumber", emptyArray<Any>(), stringDecoder)
-            batch.addRpcCall(call1)
-
-            val batchResult = httpClient.requestBatch(batch).get()
-            batchResult shouldBe false
-
-            (batch.responses[0].get() as Result<String, RpcError>).isFailure() shouldBe true
-        }
-
-        test("empty batch") {
-            mockWebServer.enqueue(createMockResponse("[]")) // Empty JSON array response
-
-            val batch = BatchRpcRequest(0)
-
-            val batchResult = httpClient.requestBatch(batch).get()
-            batchResult shouldBe true // Should complete successfully even with no requests
-        }
-
-        test("single item batch") {
-            mockWebServer.enqueue(createMockResponse(SINGLE_ITEM_BATCH_RESPONSE))
-
-            val batch = BatchRpcRequest(1)
-            val call1 = RpcCall(httpClient, "eth_blockNumber", emptyArray<Any>(), stringDecoder)
-            batch.addRpcCall(call1)
-
-            val batchResult = httpClient.requestBatch(batch).get()
-            batchResult shouldBe true
-
-            val result = batch.responses[0].get() as Result<String, RpcError>
-            result.isSuccess() shouldBe true
-            result.unwrap() shouldBe "0x1234567"
-        }
-
-        test("batch with invalid request ID in response") {
-            mockWebServer.enqueue(createMockResponse(BATCH_INVALID_ID_RESPONSE))
-
-            val batch = BatchRpcRequest(1)
-            val call1 = RpcCall(httpClient, "eth_blockNumber", emptyArray<Any>(), stringDecoder)
-            batch.addRpcCall(call1)
-
-            val batchResult = httpClient.requestBatch(batch).get()
-            batchResult shouldBe false
-
-            (batch.responses[0].get() as Result<String, RpcError>).isFailure() shouldBe true
-        }
-    }
-
-    context("Request/Response format tests") {
-        lateinit var mockWebServer: MockWebServer
-        lateinit var httpClient: HttpClient
-
-        beforeEach {
-            mockWebServer = MockWebServer()
-            mockWebServer.start()
-            httpClient = HttpClient(mockWebServer.url("").toString(), OkHttpClient())
-        }
-
-        afterEach {
-            mockWebServer.shutdown()
-        }
-
-        test("JSON-RPC request structure") {
+        test("content-Type header is set correctly") {
             mockWebServer.enqueue(createMockResponse(SUCCESSFUL_RESPONSE))
 
-            httpClient.request("eth_getBalance", arrayOf("0x1234", "latest"), stringDecoder).get()
+            httpClient.request("eth_blockNumber", emptyArray<Any>(), stringDecoder).get()
 
             val request = mockWebServer.takeRequest()
-            verifyJsonRpcRequest(request, "eth_getBalance", arrayOf("0x1234", "latest"))
+            request.getHeader("Content-Type") shouldBe "application/json"
         }
 
         test("unique request IDs are generated") {
@@ -311,21 +119,11 @@ class HttpClientTest : FunSpec({
 
             body1.get("id").asLong() shouldNotBe body2.get("id").asLong()
         }
-
-        test("content-Type header is set correctly") {
-            mockWebServer.enqueue(createMockResponse(SUCCESSFUL_RESPONSE))
-
-            httpClient.request("eth_blockNumber", emptyArray<Any>(), stringDecoder).get()
-
-            val request = mockWebServer.takeRequest()
-            request.getHeader("Content-Type") shouldBe "application/json"
-        }
     }
 
     context("Subscription tests") {
-        val httpClient = HttpClient("http://localhost:8545", OkHttpClient())
-
         test("subscription is not supported") {
+            val httpClient = HttpClient("http://localhost:8545", OkHttpClient())
             val result = httpClient.subscribe(arrayOf("newHeads"), stringDecoder).get()
 
             result.isFailure() shouldBe true
@@ -334,30 +132,10 @@ class HttpClientTest : FunSpec({
             error.message shouldContain "not supported by HTTP client"
         }
     }
-})
-
-private val stringDecoder = Function<JsonParser, String> { parser -> parser.text }
+}
 
 private fun createMockResponse(body: String, code: Int = 200): MockResponse {
     return MockResponse().setResponseCode(code).setBody(body)
-}
-
-private fun verifyJsonRpcRequest(request: RecordedRequest, expectedMethod: String, expectedParams: Array<*>) {
-    val body = Jackson.MAPPER.readTree(request.body.readUtf8())
-    body.get("jsonrpc").asText() shouldBe "2.0"
-    body.get("method").asText() shouldBe expectedMethod
-    body.get("id").isNumber shouldBe true
-
-    val params = body.get("params")
-    if (expectedParams.isEmpty()) {
-        params.isArray shouldBe true
-        params.size() shouldBe 0
-    } else {
-        params.isArray shouldBe true
-        for (i in expectedParams.indices) {
-            params.get(i).asText() shouldBe expectedParams[i].toString()
-        }
-    }
 }
 
 @Language("JSON")
@@ -379,74 +157,4 @@ private val RPC_ERROR_RESPONSE = """
         "message": "Method not found"
     }
 }
-""".trimIndent()
-
-@Language("JSON")
-private val RESPONSE_MISSING_FIELDS = """
-{
-    "jsonrpc": "2.0",
-    "id": 1
-}
-""".trimIndent()
-
-@Language("JSON")
-private val MIXED_BATCH_RESPONSE = """
-[
-    {
-        "jsonrpc": "2.0",
-        "id": 1,
-        "result": "0x1234567"
-    },
-    {
-        "jsonrpc": "2.0",
-        "id": 2,
-        "error": {
-            "code": -32601,
-            "message": "Method not found"
-        }
-    }
-]
-""".trimIndent()
-
-@Language("JSON")
-private val OUT_OF_ORDER_BATCH_RESPONSE = """
-[
-    {
-        "jsonrpc": "2.0",
-        "id": 3,
-        "result": "result3"
-    },
-    {
-        "jsonrpc": "2.0",
-        "id": 1,
-        "result": "result1"
-    },
-    {
-        "jsonrpc": "2.0",
-        "id": 2,
-        "result": "result2"
-    }
-]
-""".trimIndent()
-
-@Language("JSON")
-private val SINGLE_ITEM_BATCH_RESPONSE = """
-[
-    {
-        "jsonrpc": "2.0",
-        "id": 1,
-        "result": "0x1234567"
-    }
-]
-""".trimIndent()
-
-@Language("JSON")
-private val BATCH_INVALID_ID_RESPONSE = """
-[
-    {
-        "jsonrpc": "2.0",
-        "id": 999,
-        "result": "0x1234567"
-    }
-]
 """.trimIndent()
